fix: refactored dependency injection
DI is using channels to share messages instead of linking interfaces

+++ dns/service/service.go
@@ -2,10 +2,18 @@ package service
 
 import (
 	"context"
+	"errors"
+	"reflect"
 
-	dnsr "github.com/miekg/dns"
 	"github.com/zalgonoise/x/dns/dns"
 	"github.com/zalgonoise/x/dns/store"
+)
+
+var (
+	ErrNoAddr      = errors.New("no IP address provided")
+	ErrNoName      = errors.New("no domain name provided")
+	ErrNoType      = errors.New("no DNS record type provided")
+	ErrEmtpyRecord = errors.New("record cannot be empty")
 )
 
 type Service interface {
@@ -14,8 +22,9 @@ type Service interface {
 }
 
 type service struct {
-	dns   dns.Repository
-	store store.Repository
+	dns      dns.Repository
+	store    store.Repository
+	recordCh chan *store.Record
 }
 
 func New(dnsR dns.Repository, storeR store.Repository) *service {
@@ -26,20 +35,29 @@ func New(dnsR dns.Repository, storeR store.Repository) *service {
 		storeR = store.Unimplemented()
 	}
 
-	// link DNS server to DNS Records store
-	dnsR.Store(storeR)
-
 	return &service{
 		dns:   dnsR,
 		store: storeR,
 	}
 }
 
-func (s *service) HandleRequest(w dnsr.ResponseWriter, r *dnsr.Msg) {
-	s.dns.HandleRequest(w, r)
-}
-
 func (s *service) Start() error {
+	s.recordCh = s.dns.Link()
+
+	// route queries
+	go func() {
+		for m := range s.recordCh {
+			ctx := context.Background()
+			answer, err := s.store.GetByDomain(ctx, m)
+			if err != nil {
+				// logger
+				s.recordCh <- store.New().Build()
+			} else {
+				s.recordCh <- answer
+			}
+		}
+	}()
+
 	return s.dns.Start()
 }
 
@@ -48,12 +66,16 @@ func (s *service) Stop() error {
 }
 
 func (s *service) Reload() error {
-	return s.dns.Reload()
+	err := s.dns.Stop()
+	if err != nil {
+		return err
+	}
+	return s.Start()
 }
 
-func (s *service) Store(store store.Repository) {
-	s.dns.Store(s.store)
-}
+// func (s *service) Store(store store.Repository) {
+// 	s.dns.Store(s.store)
+// }
 
 func (s *service) Add(ctx context.Context, r ...*store.Record) error {
 	return s.store.Add(ctx, r...)
@@ -64,17 +86,39 @@ func (s *service) List(ctx context.Context) ([]*store.Record, error) {
 }
 
 func (s *service) GetByDomain(ctx context.Context, r *store.Record) (*store.Record, error) {
+	if r.Name == "" {
+		return nil, ErrNoName
+	}
+	if r.Type == "" {
+		return nil, ErrNoType
+	}
+
 	return s.store.GetByDomain(ctx, r)
 }
 
 func (s *service) GetByDest(ctx context.Context, r *store.Record) ([]*store.Record, error) {
+	if r.Addr == "" {
+		return nil, ErrNoAddr
+	}
+
 	return s.store.GetByDest(ctx, r)
 }
 
 func (s *service) Update(ctx context.Context, addr string, r *store.Record) error {
+	if addr == "" {
+		return ErrNoAddr
+	}
+	// if record is nil or empty, request is parsed as a delete operation
+	if r == nil || reflect.DeepEqual(r, &store.Record{}) {
+		return s.store.Delete(ctx, &store.Record{Addr: addr})
+	}
+
 	return s.store.Update(ctx, addr, r)
 }
 
 func (s *service) Delete(ctx context.Context, r *store.Record) error {
+	if r.Name == "" && r.Type == "" && r.Addr == "" {
+		return ErrEmtpyRecord
+	}
 	return s.store.Delete(ctx, r)
 }
